tweak: use 'ed25519' instead of 'ec' to generate keypairs
Ed25519 has shorter keys, faster crypto (constant time!), and is reliable enough
to be the standard scheme for OpenSSH.

Meanwhile I was not able to find a lot of information for the ECDSA with curve
'sect239k1' that was being used before this commit. For example, this curve is
not listed among the "safe curves" list: http://safecurves.cr.yp.to/rho.html

If we must use ECDSA with NIST curves, I propose we use the more standard
P-256 (or P-384 or even P-521) curve.